+++ src/main/java/com/ayakacraft/authlibproxyforserver/utils/NetworkUtils.java
@@ -43,6 +43,9 @@ public final class NetworkUtils {
     }
 
     public static Pair<Long, Integer> tcpingMultiple(URI uri, Proxy proxy, int times) {
+        if (times <= 0) {
+            throw new IllegalArgumentException("Not positive as expected: " + times);
+        }
         long sum = 0L;
         int i = 0;
         for (int j = 0; j < times; j++) {