+++ changelog.md
@@ -1,4 +1,7 @@
 ### General
 
-- ***Added*** command `/authproxy ping`
-- Support for Minecraft 1.21.7
+- Support for Minecraft 1.21.8
+
+### Bugfix
+
+- Incorrect average ping when some but not all packages are lost